refactor: complete walk_forward modularization + add anchored/CPCV tests
Delete monolithic walk_forward.py (927 lines) — code already lives in
walk_forward/ package. Add report.py and splitting.py modules. Add tests
for anchored_walk_forward and combinatorial_purged_cv (17/17 pass).
